# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Secure Gift Card Manager API
  version: 0.1.0
  description: >
    Initial API contract for the local MVP. The implementation must keep this
    file in sync with backend routes and QA requirement IDs.
servers:
  - url: http://localhost:3001
    description: Local API server
...
paths:
...
  /api/auth/setup:
    post:
      tags: [Auth]
      summary: Create the first account/user and wrapped DEK
      operationId: setupAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Setup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Unlock encrypted data
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Clear session and in-memory unlock state
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Logout complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatusResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/auth/change-unlock-secret:
    post:
      tags: [Auth]
      summary: Rewrap the DEK with a new unlock secret
      operationId: changeUnlockSecret
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeUnlockSecretRequest"
      responses:
        "200":
          description: Secret changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Not authenticated or locked
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated request failed CSRF, origin, or authorization checks
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: State conflict or duplicate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AuthStatusResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [setupComplete, sessionValid, dekLoaded]
          properties:
            setupComplete:
              type: boolean
            sessionValid:
              type: boolean
            dekLoaded:
              type: boolean
            user:
              $ref: "#/components/schemas/AuthUser"
            csrfToken:
              type: string
    AuthUser:
      type: object
      required: [id, accountId, role]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        role:
          type: string
          enum: [owner, admin, operator, viewer]
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
...
    SetupRequest:
      type: object
      required: [unlockSecret]
      properties:
        unlockSecret:
          type: string
          minLength: 8
        displayName:
          type: string
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required: [unlockSecret]
      properties:
        email:
          type: string
          format: email
          description: Required when more than one active user exists.
        unlockSecret:
          type: string
    ChangeUnlockSecretRequest:
      type: object
      required: [oldUnlockSecret, newUnlockSecret]
      properties:
        oldUnlockSecret:
          type: string
        newUnlockSecret:
          type: string
          minLength: 8
...
    ObjectResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, code, message]
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string